Added remaining tests to pg_imdb

+++ queng_infra/postgres/pg_imdb/src/model/integration_config/integration_config_type_create_conversion.rs
@@ -8,10 +8,10 @@ impl CreateIntegrationConfig {
     ) -> CreateIntegrationConfig {
         CreateIntegrationConfig {
             integration_id: common_integration_config.integration_id().to_owned(),
-            integration_version: common_integration_config.integration_version() as i16,
-            ims_integration_type: common_integration_config.ims_integration_type() as i16,
+            integration_version: common_integration_config.integration_version() as i32,
+            ims_integration_type: common_integration_config.ims_integration_type() as i32,
             online: common_integration_config.online(),
-            exchange_id: common_integration_config.exchange_id() as i16,
+            exchange_id: common_integration_config.exchange_id() as i32,
             integration_message_config: MessageConfig::from_integration_message_config_type(
                 common_integration_config
                     .integration_message_config()
